move calc of its_theta its_r to VSTransform constructor
otherwise, calc_coords was not getting called on the b_mode_transform
and the values were not getting initialized

another option would to be to put the its_col_cos, its_col_sin,
its_theta, its_r, its_encoder_positions, into the cxx::ImageReader,
and use as references as is done with the its_image_x, e.g.
This would be faster and would not require the std::cos, std::sin
calculations during startup when scan conversion is not required.  Also
consolidate the memory and calculations required by the scout images.