Xsd directory added and ITKVisualSonicsImageIO dll issue addressed
With these changes, the project is confirmed to build on 64 bit windows 7
running VS 9 on 32x Release mode. The xerces-c_3_0_v90.dll file needs
to be copied in or added to environment for the program to execute.
BUILD_TESTING and RDI2HTML, if on, both cause errors. However due to their
adjunct nature, are disabled for the build. XSD vs 3.2 is used for the build
and the xerces library comes with the installation.